Add AutoMapper and profiles, implement Campaign CRUD

+++ RentACar.Application/Features/BodyTypes/Commands/DeleteBodyTypeCommand.cs
@@ -1,4 +1,5 @@
-﻿using MediatR;
+﻿using AutoMapper;
+using MediatR;
 using RentACar.Application.Features.BodyTypes.Dtos;
 using RentACar.Application.Interfaces.Repositories;
 using RentACar.Domain.Entities;
@@ -15,21 +16,23 @@ namespace RentACar.Application.Features.BodyTypes.Commands
         public class DeleteBodyTypeCommandHandler : IRequestHandler<DeleteBodyTypeCommand, DeleteBodyTypeDto>
         {
             private readonly IBodyTypeRepository _bodyTypeRepository;
-            public DeleteBodyTypeCommandHandler(IBodyTypeRepository bodyTypeRepository)
+            private readonly IMapper _mapper;
+
+            public DeleteBodyTypeCommandHandler(IBodyTypeRepository bodyTypeRepository, IMapper mapper)
             {
                 _bodyTypeRepository = bodyTypeRepository;
+                _mapper = mapper;
             }
             public async Task<DeleteBodyTypeDto> Handle(DeleteBodyTypeCommand deleteBodyTypeCommand, CancellationToken cancellationToken)
             {
                 BodyType bodyType = await _bodyTypeRepository.GetByIdAsync(deleteBodyTypeCommand.Id);
+
+                if (bodyType == null)
+                    throw new Exception("BodyType not found");
+
                 await _bodyTypeRepository.DeleteAsync(bodyType);
-                DeleteBodyTypeDto deleteBodyTypeDto = new()
-                {
-                    Id = bodyType.Id,
-                    Name = bodyType.Name,
-                    DeletedDate = DateTime.UtcNow
-                };
-                return deleteBodyTypeDto;
+
+                return _mapper.Map<DeleteBodyTypeDto>(bodyType);
             }
         }
     }
